refactor(#272): split db/index.ts into domain-specific modules
Split the 2106-line monolithic db/index.ts into 12 domain modules:
- db-sessions.ts (202 lines)
- db-messages.ts (220 lines)
- db-workspaces.ts (321 lines)
- db-workspace-review.ts (278 lines)
- db-agent-profiles.ts (296 lines)
- db-artifacts.ts (167 lines)
- db-projects.ts (98 lines)
- db-linear.ts (66 lines)
- db-diff-comments.ts (66 lines)
- db-deployments.ts (103 lines)
- db-workspace-providers.ts (144 lines)
- db-workspace-provisioning.ts (107 lines)

index.ts is now 177 lines: createDb(), schema init, and re-exports.
